[WFCORE-5578] Avoid unnecessary number boxing (process-controller)

+++ host-controller/src/main/java/org/jboss/as/host/controller/HostControllerService.java
@@ -224,7 +224,7 @@ public class HostControllerService implements Service<AsyncFuture<ServiceContain
         String prettyVersion = environment.getProductConfig().getPrettyVersionString();
         //Moved to AbstractControllerService.stop()
         //processState.setStopping();
-        ServerLogger.AS_ROOT_LOGGER.serverStopped(prettyVersion, Integer.valueOf((int) (context.getElapsedTime() / 1000000L)));
+        ServerLogger.AS_ROOT_LOGGER.serverStopped(prettyVersion, (int) (context.getElapsedTime() / 1000000L));
         BootstrapListener.deleteStartupMarker(environment.getDomainTempDir());
     }
